Update and rename test.margin.js to test.marginModification.js
